Add some code for golang safeexec wrapper

+++ invoker/safeexec.go
@@ -1,15 +1,17 @@
 package invoker
 
 import (
-	"time"
-	"fmt"
+	"bufio"
 	"context"
-	"os/exec"
-	"os"
-	"path/filepath"
 	"crypto/rand"
-	"syscall"
 	"encoding/hex"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+	"syscall"
+	"time"
 )
 
 type executeFileConfig struct {
@@ -25,6 +27,7 @@ type safeexecProcessConfig struct {
 	StderrPath  string
 	ImagePath   string
 	Workdir     string
+	Command     []string
 	InputFiles  []executeFileConfig
 	OutputFiles []executeFileConfig
 }
@@ -38,37 +41,34 @@ type safeexecProcess struct {
 
 func (p *safeexecProcess) Release() error {
 	if p.cmd != nil {
-		_ = p.cmd.Process.Kill()
+		if p.cmd.Process != nil {
+			_ = p.cmd.Process.Kill()
+		}
 		_ = p.cmd.Wait()
 	}
 	_ = syscall.Rmdir(p.cgroupPath)
 	return os.RemoveAll(p.path)
 }
 
-type safeexec struct {
-	path             string
-	cgroupParentPath string
-	executionPath    string
+type safeexecProcessor struct {
+	path          string
+	executionPath string
+	cgroupPath    string
 }
 
-func (m *safeexec) Execute(ctx context.Context, config safeexecProcessConfig) (*safeexecProcess, error) {
+func (m *safeexecProcessor) Execute(ctx context.Context, config safeexecProcessConfig) (*safeexecProcess, error) {
 	process, err := m.prepareProcess()
 	if err != nil {
-		return nil, err
-	}
-	process.cgroupPath = filepath.Join(m.cgroupParentPath, process.name)
-	upperdir := filepath.Join(process.path, "upper")
-	if err := os.Mkdir(upperdir, 0777); err != nil {
-		_ = process.Release()
 		return nil, err
 	}
 	var args []string
 	args = append(args, "--time-limit", fmt.Sprint(config.TimeLimit.Milliseconds()))
 	args = append(args, "--memory-limit", fmt.Sprint(config.MemoryLimit))
-	args = append(args, "--lowerdir", config.ImagePath)
-	args = append(args, "--upperdir", upperdir)
-	args = append(args, "--cgroup-parent", m.cgroupParentPath)
-	args = append(args, "--cgroup-name", process.name)
+	args = append(args, "--overlay-lowerdir", config.ImagePath)
+	args = append(args, "--overlay-upperdir", filepath.Join(process.path, "upper"))
+	args = append(args, "--overlay-workdir", filepath.Join(process.path, "workdir"))
+	args = append(args, "--rootfs", filepath.Join(process.path, "rootfs"))
+	args = append(args, "--cgroup-path", process.cgroupPath)
 	if len(config.Workdir) > 0 {
 		args = append(args, "--workdir", config.Workdir)
 	}
@@ -81,30 +81,107 @@ func (m *safeexec) Execute(ctx context.Context, config safeexecProcessConfig) (*
 	if len(config.StderrPath) > 0 {
 		args = append(args, "--stderr", config.StderrPath)
 	}
+	args = append(args, config.Command...)
 	cmd := exec.CommandContext(ctx, m.path, args...)
 	process.cmd = cmd
 	return process, nil
 }
 
-func (m *safeexec) prepareProcess() (*safeexecProcess, error) {
+func (m *safeexecProcessor) createProcessName() (string, error) {
 	for i := 0; i < 100; i++ {
 		bytes := make([]byte, 16)
 		if _, err := rand.Read(bytes); err != nil {
-			return nil, err
+			return "", err
 		}
 		name := hex.EncodeToString(bytes)
 		path := filepath.Join(m.executionPath, name)
-		if err := os.MkdirAll(path, 0777); err != nil {
+		if err := os.MkdirAll(path, os.ModePerm); err != nil {
 			if os.IsExist(err) {
 				continue
 			}
-			return nil, err
+			return "", err
 		}
-		return &safeexecProcess{name: name, path: path}, nil
+		return name, nil
 	}
-	return nil, fmt.Errorf("cannot prepare process")
+	return "", fmt.Errorf("cannot prepare process")
 }
 
-func newSafeexec() (*safeexec, error) {
-	return &safeexec{}, nil
+func (m *safeexecProcessor) prepareProcess() (*safeexecProcess, error) {
+	name, err := m.createProcessName()
+	if err != nil {
+		return nil, err
+	}
+	path := filepath.Join(m.executionPath, name)
+	cgroupPath := filepath.Join(m.cgroupPath, name)
+	if err := syscall.Rmdir(cgroupPath); err != nil {
+		if errno, ok := err.(syscall.Errno); !ok || errno != syscall.ENOENT {
+			return nil, err
+		}
+	}
+	upperdir := filepath.Join(path, "upper")
+	if err := os.Mkdir(upperdir, os.ModePerm); err != nil {
+		_ = os.RemoveAll(path)
+		return nil, err
+	}
+	workdir := filepath.Join(path, "workdir")
+	if err := os.Mkdir(workdir, os.ModePerm); err != nil {
+		_ = os.RemoveAll(path)
+		return nil, err
+	}
+	rootfs := filepath.Join(path, "rootfs")
+	if err := os.Mkdir(rootfs, os.ModePerm); err != nil {
+		_ = os.RemoveAll(path)
+		return nil, err
+	}
+	return &safeexecProcess{
+		name:       name,
+		path:       path,
+		cgroupPath: cgroupPath,
+	}, nil
+}
+
+func newSafeexecProcessor(path, executionPath, cgroupName string) (*safeexecProcessor, error) {
+	cgroupPath, err := getCgroupParentPath()
+	if err != nil {
+		return nil, err
+	}
+	if len(cgroupName) != 0 {
+		if dir := filepath.Dir(cgroupPath); strings.HasPrefix(dir, "/sys/fs/cgroup") {
+			cgroupPath = filepath.Join(dir, cgroupName)
+		}
+	}
+	if err := os.Mkdir(cgroupPath, os.ModePerm); err != nil && !os.IsExist(err) {
+		return nil, err
+	}
+	if err := os.MkdirAll(executionPath, os.ModePerm); err != nil && !os.IsExist(err) {
+		return nil, err
+	}
+	return &safeexecProcessor{
+		path:          path,
+		executionPath: executionPath,
+		cgroupPath:    cgroupPath,
+	}, nil
+}
+
+func getCgroupParentPath() (string, error) {
+	f, err := os.Open("/proc/self/cgroup")
+	if err != nil {
+		return "", err
+	}
+	defer func() { _ = f.Close() }()
+	scanner := bufio.NewScanner(f)
+	for scanner.Scan() {
+		text := scanner.Text()
+		parts := strings.SplitN(text, ":", 3)
+		if len(parts) < 3 {
+			return "", fmt.Errorf("invalid cgroup line: %q", text)
+		}
+		if parts[1] == "" {
+			return filepath.Join("/sys/fs/cgroup", parts[2]), nil
+		}
+	}
+	if err := scanner.Err(); err != nil {
+		return "", err
+	}
+	return "", fmt.Errorf("cannot find cgroup path")
 }
